add support for zkVM
zkVM's do not have a notion of file systems. Rather than storing TSC
data in a file, this change adds a way to store the TSC in a buffer with
the idea being that the TSC becomes a part of the zkVM's guest program
at build time. By including this as a part of the guest, we save the
cost of moving all of the data into the zkVM's memory. While this means
that a new setup file will result in a new program, this is not expected
to happen frequently.

Co-authored-by: Parker Thompson <parker@risczero.com>
Co-authored-by: Austin Abell <austinabell8@gmail.com>